move getPathDetails to scripts
reduce FCP by removing script dependencies
