+++ blocks/browse-breadcrumb/browse-breadcrumb.js
@@ -1,7 +1,6 @@
 import ffetch from '../../scripts/ffetch.js';
-import { getPathDetails } from '../../scripts/language.js';
 import { fetchPlaceholders } from '../../scripts/lib-franklin.js';
-import { getEDSLink, getLink } from '../../scripts/scripts.js';
+import { getEDSLink, getLink, getPathDetails } from '../../scripts/scripts.js';
 
 export default async function decorate(block) {
   // fallback text
