+++ blocks/browse-rail/browse-rail.js
@@ -1,8 +1,7 @@
 import ffetch from '../../scripts/ffetch.js';
 import { getMetadata, fetchPlaceholders } from '../../scripts/lib-franklin.js';
 import { filterSubPages, convertToMultiMap, convertToULList, sortFirstLevelList } from './browse-rail-utils.js';
-import { getEDSLink, getLink, getProducts } from '../../scripts/scripts.js';
-import { getPathDetails } from '../../scripts/language.js';
+import { getEDSLink, getLink, getProducts, getPathDetails } from '../../scripts/scripts.js';
 
 // Utility function to toggle visibility of items
 function toggleItemVisibility(itemList, startIndex, show) {
