+++ paths.yaml
@@ -4,12 +4,26 @@ mappings:
   - /content/exlm/metadata:/metadata.json
   # spreadsheets
   - /content/exlm/redirects:/redirects.json
-  - /content/exlm/placeholders:/placeholders.json
   - /content/exlm/global/en/top-products:/en/top-products.json
   - /content/dam/exlm/franklin/config.json:/.helix/config.json
   # force load some paths as resource from codebus in the editor (sw.js)
   - /content/exlm.resource/fragments/:/fragments/
   - /content/exlm.resource/docs/:/docs/
+
+  # placeholders
+  - /content/exlm/global/en/placeholders:/placeholders.json
+  - /content/exlm/global/en/placeholders:/en/placeholders.json
+  - /content/exlm/global/fr/placeholders:/fr/placeholders.json
+  - /content/exlm/global/de/placeholders:/de/placeholders.json
+  - /content/exlm/global/es/placeholders:/es/placeholders.json
+  - /content/exlm/global/it/placeholders:/it/placeholders.json
+  - /content/exlm/global/nl/placeholders:/nl/placeholders.json
+  - /content/exlm/global/pt-br/placeholders:/pt-br/placeholders.json
+  - /content/exlm/global/ko/placeholders:/ko/placeholders.json
+  - /content/exlm/global/ja/placeholders:/ja/placeholders.json
+  - /content/exlm/global/sv/placeholders:/sv/placeholders.json
+  - /content/exlm/global/zh-hans/placeholders:/zh-hans/placeholders.json
+  - /content/exlm/global/zh-hant/placeholders:/zh-hant/placeholders.json
 
   # browse inidices
   - /content/exlm.resource/en/browse-index.json:/en/browse-index.json
